+++ packages/orm/lib/src/runtime/json_protocol/serialize.dart
@@ -38,13 +38,21 @@ extension ModelActionToJsonQueryAction on dmmf.ModelAction {
   }
 }
 
-extension on JsonQueryAction {
-  String toApiFunctionName() {
-    if (name.endsWith('One')) return name.substring(0, name.length - 3);
+// TODO: What's this?
+//
+// I wrote it myself, but I forgot why I wrote it and why it is not used.
+//
+// Maybe I need to re-verify the Prisma JSON protocol
+//
+// -----------------------------------
+//
+// extension on JsonQueryAction {
+//   String toApiFunctionName() {
+//     if (name.endsWith('One')) return name.substring(0, name.length - 3);
 
-    return name;
-  }
-}
+//     return name;
+//   }
+// }
 
 extension<K, V> on Map<K, V> {
   Map<K, V> withoutKeys(Iterable<K> keys) =>